refactor: Move GameService from API
