+++ Town-Of-Salem/src/main/java/com/github/raresp/proiectip/TownOfSalem/utils/GameRunner.java
@@ -1,6 +1,6 @@
 package com.github.raresp.proiectip.TownOfSalem.utils;
 
-import com.github.raresp.proiectip.TownOfSalem.API.GameService;
+import com.github.raresp.proiectip.TownOfSalem.repositories.GameService;
 import com.github.raresp.proiectip.TownOfSalem.exceptions.GameNotFoundException;
 import com.github.raresp.proiectip.TownOfSalem.models.Game;
 import com.github.raresp.proiectip.TownOfSalem.models.GameState;
@@ -8,9 +8,7 @@ import com.github.raresp.proiectip.TownOfSalem.models.TurnInteractions;
 import com.github.raresp.proiectip.TownOfSalem.models.VotingSession;
 import com.github.raresp.proiectip.TownOfSalem.models.characters.Character;
 import com.github.raresp.proiectip.TownOfSalem.models.characters.SelectionSession;
-import com.github.raresp.proiectip.TownOfSalem.models.characters.TownCharacters.Sheriff;
 import com.github.raresp.proiectip.TownOfSalem.repositories.GameRepository;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.concurrent.ConcurrentTaskScheduler;
@@ -18,9 +16,6 @@ import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 
 import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.*;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
